+++ NEWS.md
@@ -5,8 +5,10 @@ The current version of the DLMtool package is available for download from [CRAN]
 - improve convergence diagnostic in `Converge`
 - modified CAL generation to deal with dome-shaped selectivity better 
 - improved help documentation for functions and MPs by combining and linking similiar functions
-- add ImputeLH - impute correlated life-history parameters - in development
-- option for variable management interval - to add to FAQ
+- makeMeanMP - added - TODO Describe
+- option for variable management interval - TODO add to FAQ
+- added LHdatabase from FishLife and functions to predict life-history parameters 
+- add LH2OM and  predictLH - predict correlated life-history parameters - TODO add to user guide/FAQ
 
 ## DLMtool 5.1.3 
 - fix Replace function for new OM slots